Add filtering by container labels
This commit adds the ability to filter containers by their labels in the
Docker integration. The new functionality allows users to specify
container labels in the configuration, instead of container names, to
allow for dynamic container name resolution.
This is particularly useful in environments where container names may
change frequently or are not defined in advance.

Fixes #193

Signed-off-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>